web_working/requirements.txt
flask==3.0.2
flask-cors==4.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
numpy==1.24.3
tensorflow-macos==2.13.0
tensorflow-metal==1.1.0
scikit-learn==1.0.2
pyyaml==6.0.1
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.0
twilio==8.10.0
messagebird==1.2.0
requests==2.31.0